Add functional scraper for historical goodwe data
The scraper is intended to be used as a context-manager
Upon creation a request for the main page of goodwe and
a request to log in are made

After creation a request_export call can be made for a date, which will be
truncated to year and month. This will result in the creation of a export
file on the goodwe server. The return value of request_export contains the
info required to download the data.

To download the file pass the result to download_data